# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "COTC DB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Liz",
            "email": "j2qgehn84@mozmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/travellers/{id}": {
            "get": {
                "description": "get traveller information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Traveller"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Influence": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.Traveller": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "influence": {
                    "$ref": "#/definitions/domain.Influence"
                },
                "influence_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "integer"
                }
            }
        },
        "rest.StandardAPIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "metadata": {}
            }
        }
    }
}